Add /delwarp, rename /location to /warp

+++ yawa/src/main/kotlin/work/gavenda/yawa/essentials/PlayerDb.kt
@@ -1,3 +1,21 @@
+/*
+ * Yawa - All in one plugin for my personally deployed Vanilla SMP servers
+ *
+ * Copyright (c) 2022 Gavenda <gavenda@disroot.org>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
 package work.gavenda.yawa.essentials
 
 import org.jetbrains.exposed.dao.UUIDEntity
@@ -20,9 +38,9 @@ class PlayerHomeDb(uuid: EntityID<UUID>) : UUIDEntity(uuid) {
 
 object PlayerHomeSchema : UUIDTable("yawa_player_home", "uuid") {
     val world = uuid("world_uuid")
-    val x = integer("x")
-    val y = integer("y")
-    val z = integer("z")
+    val x = double("x")
+    val y = double("y")
+    val z = double("z")
 }
 
 class PlayerLocationDb(uuid: EntityID<UUID>) : UUIDEntity(uuid) {
@@ -40,7 +58,7 @@ object PlayerLocationSchema : UUIDTable("yawa_player_location", "uuid") {
     val playerUuid = uuid("player_uuid")
     val name = varchar("name", 64)
     val world = uuid("world_uuid")
-    val x = integer("x")
-    val y = integer("y")
-    val z = integer("z")
+    val x = double("x")
+    val y = double("y")
+    val z = double("z")
 }
